Add MutexContext
Its possible for the single instance manager to throw an exception on
shutdown, when it was disposed on an other thread, the the original
`StartApplication` was made. We don't want to force the user to keep
track of the correct threads themself, so we pass all calls to the mutex
to a separate thread. This however imposes an other issue for special
cases like our unut tests.

Using a dedicated thread for locking is no problem, as long as one
process only ever attempts to lock the mutex once. But when we try to
lock multiple times from from within the same application, like we do in
our unit tests, then we run into the situation, that we lock twice on
the smae threads, which will succeed both times, since a mutex is
reentrant. This however is not thre result we want, as we still want the
second call to fail and signal, that there is already a running
instance. To circumvent this issue, we track how often the lock has been
taken, and release it immedialty, when the count is greater than 1.
After that we can pretent, the lock failed, and the rest works fine
again